fixed failing compilation with high_precision enabled

+++ examples/support/examples_renderer.rs
@@ -66,9 +66,9 @@ impl ExamplesRenderer {
         ];
 
         for contact in contacts.iter() {
-            model_matrix[3][0] = contact.x;
-            model_matrix[3][1] = contact.y;
-            model_matrix[3][2] = contact.z;
+            model_matrix[3][0] = contact.x as f32;
+            model_matrix[3][1] = contact.y as f32;
+            model_matrix[3][2] = contact.z as f32;
 
             surface.draw(
                 (&self.cube.vertices, &self.cube.normals),
@@ -107,17 +107,17 @@ impl ExamplesRenderer {
         let r32 = 2.0*quat.j*quat.k + 2.0*quat.r*quat.i;
         let r33 = 1.0 - 2.0*quat.i*quat.i - 2.0*quat.j*quat.j;
 
-        let m11 = instance.scale.0 * r11;
-        let m12 = instance.scale.0 * r12;
-        let m13 = instance.scale.0 * r13;
+        let m11 = instance.scale.0 * r11 as f32;
+        let m12 = instance.scale.0 * r12 as f32;
+        let m13 = instance.scale.0 * r13 as f32;
         let m14 = instance.scale.0 * transform.translation().x as f32;
-        let m21 = instance.scale.1 * r21;
-        let m22 = instance.scale.1 * r22;
-        let m23 = instance.scale.1 * r23;
+        let m21 = instance.scale.1 * r21 as f32;
+        let m22 = instance.scale.1 * r22 as f32;
+        let m23 = instance.scale.1 * r23 as f32;
         let m24 = instance.scale.1 * transform.translation().y as f32;
-        let m31 = instance.scale.2 * r31;
-        let m32 = instance.scale.2 * r32;
-        let m33 = instance.scale.2 * r33;
+        let m31 = instance.scale.2 * r31 as f32;
+        let m32 = instance.scale.2 * r32 as f32;
+        let m33 = instance.scale.2 * r33 as f32;
         let m34 = instance.scale.2 * transform.translation().z as f32;
 
         let model_matrix: [[f32; 4]; 4] = [
